// front/src/app/models/korisnik.ts
export class Korisnik {
    korisnickoIme: string = "";
    lozinka: string = "";
    ime: string = "";
    prezime: string = "";
    pol: string = "";
    adresa: string = "";
    kontaktTelefon: string = "";
    email: string = "";
    profilnaSlika: string = "";
    brojKreditneKartice: string = "";
    status: string = "";
    tip: string = "";
}